requirements.txt
flask
flask-restful
flask-jwt-extended
flask_sqlalchemy
flask_cors
pymysql
PyCryptodome
Werkzeug
flask_bcrypt
xgboost==1.5.0rc1